-Arreglo: error delete, muevo inputs
- Arreglo error de los metodos Delete de los archivos, ejemplo. borraba cliente, pero en cuentas.txt y cuentasDeClientes.txt seguia apareciendo, hago que cuando se elimine algo de un archivo todo lo relacionado en otros se elimine tambien,
-Arreglo cuando el usuario ingrese un DNI pueda ingresar con cualquier CVU que este cargado en Cuentas.txt, ahora valido si el DNI tiene la relacion con el CVU dado con la  funcion "relacionDniYCbu".
- Muevo todas los inputs de CVU y DNI, ingresar dinero a presentacion

+++ src/main/java/ar/edu/utn/frbb/tup/persistence/ClienteDao.java
@@ -18,13 +18,11 @@ public class ClienteDao {
         try {
             //Me fijo si el archivo existe
             File archivo = new File(RUTA_ARCHIVO);
-            boolean existe = archivo.exists();
 
-            //Si no existe, lo creo y guardo el Encabezado para saber el orden de los datos
-            FileWriter fileWriter = new FileWriter(RUTA_ARCHIVO, true);
-            writer = new PrintWriter(fileWriter);
-
-            if (!existe) {
+            if (!archivo.exists()) {
+                //Si no existe, lo creo y guardo el Encabezado para saber el orden de los datos
+                FileWriter fileWriter = new FileWriter(RUTA_ARCHIVO, true);
+                writer = new PrintWriter(fileWriter);
                 writer.println("DNI, Nombre, Apellido, Direccion, Fecha nacimiento, Mail, Banco, Tipo Persona, Fecha alta");
             }
 
@@ -65,9 +63,7 @@ public class ClienteDao {
             while ((linea = reader.readLine()) != null) { //Condicion para que lea el archivo hasta el final y lo guarde en la variable linea
                 String[] datos = linea.split(",");
 
-                if (Long.parseLong(datos[0]) == dni){ //Si encuentro el dni que ingreso el cliente, no lo agrego en el contenido por ende lo elimino
-                    continue;
-                } else {
+                if (Long.parseLong(datos[0]) != dni){ //Voy agregando todas las lineas del Archivo excepto las lineas que tengo que eliminar con el DNI dado
                     contenido.append(linea).append("\n"); //Agrego la linea al contenido
                 }
             }
